Make some changes to security

+++ src/main/java/com/gmail/at/ivanehreshi/epam/touragency/controller/ReviewController.java
@@ -2,6 +2,7 @@ package com.gmail.at.ivanehreshi.epam.touragency.controller;
 
 import com.gmail.at.ivanehreshi.epam.touragency.dispatcher.*;
 import com.gmail.at.ivanehreshi.epam.touragency.domain.*;
+import com.gmail.at.ivanehreshi.epam.touragency.security.*;
 import com.gmail.at.ivanehreshi.epam.touragency.service.*;
 import com.gmail.at.ivanehreshi.epam.touragency.util.*;
 
@@ -25,6 +26,12 @@ public class ReviewController extends Controller {
     @Override
     public void delete(RequestService reqService) {
         Long id = reqService.getLong("id").get();
+        Review review = reviewService.read(id);
+        User user = reqService.getUser().get();
+        if (!review.getAuthor().getId().equals(user.getId())
+                && !reqService.getRequest().isUserInRole(Role.TOUR_AGENT.name())) {
+            throw new AccessDeniedException();
+        }
         reviewService.delete(id);
         reqService.redirect(reqService.getRequest().getHeader("Referer"));
     }
